Added the following tests to file test_harvester_global_soilm4.py test_gridcell_area_conservation() test_global_mean_values() test_global_mean_values2() test_gridcell_variance() test_gridcell_min_max() Deleted bfg soilt4 files from the test directory.  They are not used in the test of this harvester.
